+++ package.json
@@ -26,10 +26,8 @@
 		"@nextui-org/theme": "2.2.0",
 		"@react-aria/ssr": "^3.9.2",
 		"@react-aria/visually-hidden": "^3.8.10",
-		"autoprefixer": "10.4.19",
 		"classnames": "^2.5.1",
 		"clsx": "^2.0.0",
-		"eslint": "8.48.0",
 		"eslint-config-next": "^14.0.3",
 		"framer-motion": "^11.1.8",
 		"intl-messageformat": "^10.5.0",
@@ -38,15 +36,12 @@
 		"next-auth": "^4.24.5",
 		"next-swagger-doc": "^0.4.0",
 		"next-themes": "^0.2.1",
-		"postcss": "8.4.38",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-grid-layout": "^1.4.4",
 		"react-stately": "^3.31.0",
 		"swagger-ui-react": "^5.17.10",
-		"tailwind-variants": "^0.1.20",
-		"tailwindcss": "3.4.3",
-		"typescript": "5.0.4"
+		"tailwind-variants": "^0.1.20"
 	},
 	"devDependencies": {
 		"@types/node": "20.9.3",
